+++ internal/lookoutingesterv2/configuration/types.go
@@ -3,6 +3,7 @@ package configuration
 import (
 	"time"
 
+	commonconfig "github.com/armadaproject/armada/internal/common/config"
 	profilingconfig "github.com/armadaproject/armada/internal/common/profiling/configuration"
 	"github.com/armadaproject/armada/internal/server/configuration"
 )
@@ -13,7 +14,7 @@ type LookoutIngesterV2Configuration struct {
 	// Metrics configuration
 	MetricsPort uint16
 	// General Pulsar configuration
-	Pulsar configuration.PulsarConfig
+	Pulsar commonconfig.PulsarConfig
 	// Pulsar subscription name
 	SubscriptionName string
 	// Size in bytes above which job specs will be compressed when inserting in the database
