+++ internal/scheduleringester/config.go
@@ -3,6 +3,7 @@ package scheduleringester
 import (
 	"time"
 
+	commonconfig "github.com/armadaproject/armada/internal/common/config"
 	profilingconfig "github.com/armadaproject/armada/internal/common/profiling/configuration"
 	"github.com/armadaproject/armada/internal/server/configuration"
 )
@@ -13,7 +14,7 @@ type Configuration struct {
 	// Metrics Port
 	MetricsPort uint16
 	// General Pulsar configuration
-	Pulsar configuration.PulsarConfig
+	Pulsar commonconfig.PulsarConfig
 	// Pulsar subscription name
 	SubscriptionName string
 	// Number of event messages that will be batched together before being inserted into the database
